Update student using modal

+++ frontend-development/vue/crud-operations/src/components/body.vue
@@ -1,7 +1,8 @@
 <template>
   <div class="col-9 offset-1">
     <addStudent @addbtnClicked="addNewStudent" />
-    <table class="table table-stripped table-bordered text-center">
+
+    <table class="table table-striped table-bordered text-center">
       <thead>
         <tr>
           <th>ID</th>
@@ -15,8 +16,12 @@
           <td>{{ student.id }}</td>
           <td>{{ student.name }}</td>
           <td>{{ student.city }}</td>
-          <td><i class="fa-solid fa-trash mx-2" @click="removeStudent(index)"></i> | <i
-              class="fa-solid fa-pen-to-square mx-2"></i></td>
+          <td>
+            <i class="fa-solid fa-trash mx-2" @click.ctrl="removeStudent(index)"></i> |
+            <i class="fa-solid fa-pen-to-square mx-2" data-bs-toggle="modal" data-bs-target="#editModal"
+              @click="setSelectedStudent(student)">
+            </i>
+          </td>
         </tr>
       </tbody>
       <tfoot>
@@ -25,39 +30,79 @@
         </tr>
       </tfoot>
     </table>
+
+    <div class="modal fade" id="editModal" data-bs-keyboard="false" data-bs-backdrop="static">
+      <div class="modal-dialog modal-dialog-centered">
+        <div class="modal-content">
+          <div class="modal-header">
+            <h4 class="text-center">Edit Student</h4>
+          </div>
+          <div class="modal-body">
+            <input type="text" class="form-control my-2" v-model="selectedStudent.id" disabled>
+            <input type="text" class="form-control my-2" v-model="selectedStudent.name">
+            <input type="text" class="form-control my-2" v-model="selectedStudent.city">
+          </div>
+          <div class="modal-footer">
+            <button class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
+            <button class="btn btn-primary" data-bs-dismiss="modal" @click="updateStudent">Save Changes</button>
+          </div>
+        </div>
+      </div>
+    </div>
   </div>
 </template>
 
 <script>
 import addStudent from "./addStudent.vue";
+
 export default {
   components: {
     addStudent
   },
   data: () => ({
     students: [],
+    selectedStudent: { id: "", name: "", city: "" }
   }),
   async created() {
     let res = await fetch("http://localhost:5000/students");
     this.students = await res.json();
   },
   methods: {
+    setSelectedStudent(student) {
+      this.selectedStudent = { ...student }; // Clone to avoid modifying directly
+    },
+    async updateStudent() {
+      const index = this.students.findIndex(s => s.id === this.selectedStudent.id);
+      if (index !== -1) {
+        await fetch(`http://localhost:5000/students/${this.selectedStudent.id}`, {
+          method: "PUT",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify(this.selectedStudent)
+        });
+
+        this.students[index] = { ...this.selectedStudent };
+      }
+      //document.querySelector("#editModal .btn-secondary").click();
+    },
     async addNewStudent(student) {
-      let data = { id: this.students[this.students.length - 1].id + 1, name: student.name, city: student.city }
+      let newId = this.students.length ? this.students[this.students.length - 1].id + 1 : 1;
+      let data = { id: newId, name: student.name, city: student.city };
+
       await fetch("http://localhost:5000/students", {
         method: "POST",
-        headers: {
-          "Content-type": "application/json",
-        },
+        headers: { "Content-type": "application/json" },
         body: JSON.stringify(data)
-      })
-      this.students.push(data)
+      });
+
+      this.students.push(data);
     },
     async removeStudent(index) {
-      await fetch(`http://localhost:5000/students/${this.students[index].id}`, {
-        method: "DELETE"
-      })
-      this.students.splice(index, 1);
+      if (confirm("Are you sure you want to delete this student?")) {
+        await fetch(`http://localhost:5000/students/${this.students[index].id}`, {
+          method: "DELETE"
+        });
+        this.students.splice(index, 1);
+      }
     }
   }
 };
